requirements.txt
torch>=1.10.0
torch-geometric>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
PyYAML>=5.4.0
matplotlib>=3.3.0
tensorboard>=2.7.0
wandb>=0.12.0
tqdm>=4.60.0 